Fetch polyfill, Cypress ids and classes, myHttp quick refactoring
