+++ package.json
@@ -103,7 +103,6 @@
     "html-webpack-plugin": "^2.7.1",
     "imports-loader": "^0.6.5",
     "ip": "^1.1.2",
-    "isomorphic-fetch": "^2.2.1",
     "jquery": "^3.0.0",
     "json-loader": "^0.5.4",
     "koa": "^2.0.0-alpha.3",
@@ -145,6 +144,7 @@
     "webpack": "^1.12.14",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-hot-middleware": "^2.6.0",
+    "whatwg-fetch": "^2.0.4",
     "xml2js": "0.4.16",
     "yargs": "^4.0.0"
   },
